README: remove badges
I'd like to switch the CI to GitHub Actions at some point.
In the meantime, I don't care about the Travis CI build.

And I certainly don't care about the code coverage; it's hard to test a
plugin like this that relies on a remote RESTful API. Of course I could
mock it, but there are only so many hours in the day.